movies_admin/requirements/base.txt
django==3.1
django-model-utils==4.0.0
django-environ==0.4.5
psycopg2-binary==2.9.1
numpy==1.21.1
pandas==1.3.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Pillow==8.3.1
sorl-thumbnail==12.7.0
djangorestframework==3.12.4
